Split Python high-level API into multiple files and made loading of the submodules dynamic
