+++ pynest/setup.py.in
@@ -28,7 +28,7 @@ setup(
     author='The NEST Initiative',
     url='http://www.nest-simulator.org',
     license='GPLv2+',
-    packages=['nest', 'nest.tests'],
+    packages=['nest', 'nest.tests', 'nest.lib'],
     package_dir={'nest': '@PKGSRCDIR@/pynest/nest'},
     package_data={'nest': ['pynest-init.sli']},
 )
